Improve security and UX based on code review feedback

+++ examples/symfony/src/Controller/UserController.php
@@ -71,10 +71,13 @@ class UserController extends AbstractController
     }
 
     #[Route('/users/{id}/delete', name: 'user_delete', methods: ['POST'])]
-    public function delete(User $user, EntityManagerInterface $em): Response
+    public function delete(User $user, Request $request, EntityManagerInterface $em): Response
     {
-        $em->remove($user);
-        $em->flush();
+        // Validate CSRF token
+        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+            $em->remove($user);
+            $em->flush();
+        }
         
         return $this->redirectToRoute('user_list');
     }
